Split filesystem builtins from file/stream
